+++ sources/Google.Solutions.IapDesktop.Extensions.Session.Test/Protocol/App/TestSsmsClient.cs
@@ -21,6 +21,7 @@
 
 
 using Google.Solutions.IapDesktop.Core.ClientModel;
+using Google.Solutions.IapDesktop.Core.ClientModel.Protocol;
 using Google.Solutions.IapDesktop.Core.ClientModel.Transport;
 using Google.Solutions.IapDesktop.Extensions.Session.Protocol.App;
 using Moq;
