+++ sources/Google.Solutions.IapDesktop.Extensions.Session/Protocol/App/SsmsClient.cs
@@ -19,7 +19,7 @@
 // under the License.
 //
 
-using Google.Solutions.IapDesktop.Core.ClientModel;
+using Google.Solutions.IapDesktop.Core.ClientModel.Protocol;
 using Google.Solutions.IapDesktop.Core.ClientModel.Transport;
 using System.Drawing;
